Finally able to add 2 features
I was finally able to add the collision box and made it so they do collide, it simply closes the program. Quite aggressive, but couldn't think of anything else.

Next feature was 'respawning' the moving square if it goes offscreen. So far it only manages to spawn back randomly within the x axis. Might have to tweak some stuff with the y axis.